Refactored location of filter API and added filteredSnapshot to the session
